Initial commit
Project simply consists of a camera and quad with a material and shader that draws the test image as a array of pixel.png instances.

It does that by sampling the color of the nearest pixel and its 4 neighbors, then mutliplying them by the corresponding points on the pixel texture, summing the results. Sampling surrounding pixels is necessary to support the potentially non-local effect of overlapping pixels (this depends on pixel shape and size).