Color divergence
Red, green, and blue signals can now fail to converge towards the edges of the image. This produces an effect like chromatic aberration.

This is applied as the first transformation to the input signal, so the effect is discretized to the virtual pixels, resulting in a more complex, and more accurate, effect. Essentially, I'm trying to mimic what would happen if the electron guns were mis-calibrated and thus each of the guns partially hit different pixels.