Scanline blur, box blur scaling
Added an effect intended to simulate a poor signal, resulting in horizontal lines of pixels sharing a small amount of their neighbor's color.

I did this by blurring the input with a horizontal box blur. I also added the ability to scale the box blur kernel because otherwise even the smallest kernel was too strong.