Pixel response time
The red, green, and blue components of each pixels now take a short time to fully reach the signal value. This behavior is assymetric; the pixel can turn on at a different rate than it turns off.

This is accomplished by lerping the last activation level towards the signal at a rate proportional to time since last frame divided by the response time. This is not completely framerate independent, but its seems to work well enough for now.

I also added an animation to the input signal to test out temporal effects like this.