Blank screen texture
Added a tiling texture of what a pixel may look like with the screen off and the lights on.

I added it to the Synthesize step, along with a handle to make it brighter or darker. This effect directly reduces contrast, so its best used very subtly. I think it is very important however, because otherwise the contrast with black feels way too good when compared to all the other visual artifacts.