Texture coverage of monitor edge
The 0-1 range of the UV coordinates on the monitor model now cover the whole screen.

To bring the virtual pixels back in from the edge, I scale and bias the texture coordinates based on a relative image size knob in Pixelize.shader.

The result is the edge of the screen is no longer black, instead recieving the glow of nearby pixels as well as the subtle grey of the surface being lit from ambient light in the room. It also removes the harsh edge to pixels on the border.