Scene.cs, Prefabs.cs, Style.cs
Added files for accessing and controlling global info.

Scene stores references to things like _the_ Taskbar, _the_ WindowContainer, _the_ UICamera, etc.

Prefabs are accessible through Scene and are where references to prefabs that must be instantiated in a script are placed.

Style contains information that controls the visualization that does not naturally belong to an editor and code accessible singleton. So, _the_ Taskbar has a height that can be changed in the editor and can be accessed programmatically through Scene, thus it is not generally appropriate for Style. Whereas the TaskbarElement prefab can have a default width set in the editor, but cannot easily be accessed programmatically, so the default width is stored in Style. The values in Style are not intended to be  redundant; they should be authoritative and if not the primary source, connect to the primary source. Something like the default margin size of visual elements is a great fit because it belongs to everyone and no one.

Prompting these changes was mainly the feature of being able to detect which visual elements were being pointed at by the virtual cursor, and which element of those was topmost. Some other changes were prompted by the new systems. The net effect is a tangled commit, so a lot of stuff is included here.

InputUtility.IsPointedAt()
InputUtility.IsTouched()
UIElement parent class
Window now interacts implements its state variables
  IsOpened
  IsMinimized
  IsMaximized
Taskbar now spawns TaskbarElements in for open Windows
Clicking TaskbarElements toggles matching Window's minimization state
Can now drag Window by title bar