InputUtility.TouchedCanvasElement bugfix
Before, it was assumed that GraphicRaycaster raycasts would always hit the cursor first so we just used the second hit for determining topmost element.

But this wasn't always true, so now I've made the Cursor not block raycasts at all so the first result of the raycast is always the correct one.