Page, ProfilePage, NavigationButton
Added a new base class/prefab for the content of Windows called "Page".

ProfilePage is the first subclass and implements a simple UI that displays information about a subject (a proper noun, like a person, ship, or company) and provides links to dig deeper. These links are represented by NavigationButtons, which have a text title and a small image.

In support of this,

1) Window now has the ability to Open() pages, as well as keep a history of them.
2) Window has some defaults as part of its prefab to aid in previewing pages in the prefab editor.
3) Added GridContainer that displays contents in an organized grid.