SystemMap.cs, SystemMapObject.cs
Added new prefabs and classes to implement an interactive map of the solar system.

The map automatically sizes the icons and distances logarithmically to maximize usage of visual space and limit overlap of icons. Clicking on an icon will center the view on that body.

Added Clock.cs as well. This class controls the game's imaginary date and time it takes place in. Its speed can be controlled through the editor. The planet's motions hook into this clock.

Logarithmic distances are very helpful for making the inner planets not pile up on eachother (and the sun). However, it remains to be seen how well this works with transit visualization.