Crafts, Parts, Hohmann Transfers, and More!
What began as an attempt to add realistic hohmann transfers escalated into a commit that has decimated countless hours of my life.

Essentially, I added in Craft.cs because I wanted to actually see the hohmann transfer implementation in action. That in turn led to the creation of Station and Visitable, and subsequentally a massive undertaking to hook all that stuff into the SystemMap UI. Along the way, I created Craft.Part to build out the Craft's behavior and data in a component based way, resulting in Engine, Navigation, FluidContainer, Life Support, Bridge, etc. Finally, I needed reaction mass to use during acceleration, so Item.cs came into being, resulting in LiquidOxygen and LiquidHydrogen, as well as supporting components PhysicalItem, LiquidItem, GasItem, and... well... PerishableItem and Person, because someone needed to be manning that bridge.