ScrollBar
ScrollBar is what it sounds like. This one is implemented to work fairly independently; you simply provide it with two nested RectTransforms, and if the parent is smaller than the child, the ScrollBar appears and can be dragged to the adjust what part of the child appears within the parent's bounds.