Item.cs additional data
Item now has a description, short name, units, a "qualifier" for details like color or size, and large and small sprites.

It also now has a function property that returns a string to describe the item's quantity. Interacting with this are the the scripts MassItemQuantityStringController and VolumeItemQuantityStringController. These produce short strings that describe the items mass or volume and are attached by default to PhysicalItems and Liquid/GasItems, respectively.

I filled in this information for two example items in the example ship.