RefuelingPage
Added a Page for the Craft refueling use case.

Allows use to buy fuel in arbitrary or meaningful amounts; meaningful amounts being "enough for first transfer", "enough for first two transfers", or 100% filled. Arbitrary amounts can be selected by clicking on a fuel meter displayed on this Page.

There are no sellers of fuel yet, so the price of fuel is just a float field currently. In addition, users do not currently have money, so the payment button is not implemented.