User, Bank, Market, Arbitrageur, Storage, Mine
I found it very difficult to add trade into the game in tiny steps. So instead, here's a very large initial step.

The main addition here is the concept of the User. In this context, it refers to the account a person or organization uses to access the in-world monolithic service that provides just about everything, including shopping, communication, running a business, travel planning, etc.

From the perspective of the real software, User is a type that defines the username, password, and ownership of things like bank accounts and Crafts. User can have User.Scripts which implement behavior/data. This is initially useful for

1) the User.Script Bank, which is what it sounds like, a wealthy User that provides BankAccounts and (in the future) loans and other financial tools.

2) The beginnings of AI; the first example is the User.Script Arbitrageur. This script automatically uses any crafts and cash at its disposal to make income through trade. It finds the most valuable destination and collection of goods for each Craft it owns and then performs the transactions necessary to buy those goods, sends the ship to the destination and sells the goods there.

Introduced to support Arbitrageur are the classes Market and Storage.

Markets allow the posting of SaleOffers and PurchaseOffers, as well as providing the function of simplifying the use of these offers when performing an immediate sale or purchase. F.e. There's 100 widgets offered for sale at a Station A  from 3 different sellers. You ask the market to purchase you 70. The Market buys from the lowest priced seller until their stock runs out, then the next in line until your order has been fulfilled.

Finally, Storage allows for a similar function of being an intermediary for an underlying panoply. In this case, however, the underlying collection is ItemContainers.