People.cs
Person has become People and its subcomponent Individual.

People is a User and Item that attempts to spend its money in order to maximize the positive change in its quantity.

People has Needs like a need for water, food, space, entertainment, safety, etc. Each need's fulfillment controls the People's overall growth rate.

To maximize this growth rate, People look at the effect a marginal purchase would make; e.g. compare the effect of a small amount of water to a small amount of food to a small amount of extra living space. It continues this process until the marginal value of savings is greater than the marginal value of an additional purchase.

Or it would if it were finished. Right now it can only judge purchases of different items.

Also TODO is making needs more abstract than a need for a particular item; "Food" shouldn't be an item necessarily but something an item provides. So, for example, a cheeseburger would provide x units of food and y units of water. Then you can have a richer retail market because People can compare different products to eachother. Bias (irrational measurement of growth rate) is another part of this, but it will probably have to wait until Retailer is implemented.

I'm changing my development focus away from Users for now and so I'm just wrapping up this feature.