StationVisualization.cs
StationVisualization.cs and its related classes and prefabs RingVisualization, WingVisualization, and StationVisualizationCamera integrate with the logical representation of Station to display the station as a 3d object.

Station's use of Rooms thus far was quite informal and mostly just data (Station's had Rooms because I dragged a Room prefab into them, not because of any game rules to defined where/why/how they exist). With this update, it now  has a concept of Station.Modules (i.e. habitat module, engine module, defense modules, etc) of which the first child class is Ring. Ring represents and spinning human habitat divided into floors, wings, and rooms. It currently generates a sample set of of these structures, which is read by StationVisualization to display them.

StationVisualization can be seen within the game through the new window StationViewer, which displays the output of StationVisualization's virtual camera.