StationToolbox, Subwindow
Added the concept of a "Subwindow"; a Window that belongs to another window. Subwindows cannot be minimized, maximized, or closed and always stay on top of their parent window.

Subwindows were added to support the main feature of this commit, which is the StationToolbox. This Subwindow has buttons that allow the user to add and remove floors and wings to a Ring.