MarketPage, NumericInput, PercentageGizmo
Added a user interface to access a Market.

This interface allows users to browse the items available in a Market and buy/sell/make an offer on these items.

It reuses code from InventoryPage by pulling out the common functionality and putting it in the new class ItemCollectionPanel.

It adds two new UI primitives NumericInput and PercentageGizmo.

NumericTextInput and its respective prefab TextInput (which generically handles stylized text input) allows for simple programmatic access to a TextInput while treating it as numeric information.

PercentageGizmo allows for visual and tactile selection of a fractional value. The visualization displays a "pie" and 0-100 text reflecting what proportion the user has selected, and by clicking and dragging on the visualization, users can change this value at a rate relative to the current value. Meaning, when you start dragging from zero, one unit of drag may equal a change of 0.05% / 100%, because the current value is very small. But by the time you've reached 50%, one unit of drag may equal 5% / 100%. This balances the need for fine adjustments and the desire to mimize the time it takes to reach higher fractions. Works best if the driven values are zero-based and the maximum value is not many orders of magnitude larger than the smallest meaningful non-zero value.

Market needed updating because it still required "real" items in many cases as opposed to identifiers and/or "samples" (mainly for physical considerations like volume). the changes I went with are that now Offers require a sample, and may have any quantity specfied as the amount offered, but this is automatically reduced to an actionable quantity when querying quantity through .AvailableSupply/Demand (e.g. If you offer 1000 widgets but only have 400, then AvailableSupply == 400). Market now exclusively uses unique identifiers (Item.Name) for everything, which makes it simpler to interact the with the marketplace, as well as making intuitive sense; apart from the actual transaction, the market's purpose is to simply facilitate communication between buyers and sellers, actual items need not enter the picture unless the market was storing the items, which I've decided against because if the user wants to buy temporary, scalable storage for items they are offering, then that will be available, but if they choose to use any other storage assets they have, they shouldn't be prevented from doing so.