Scene.The.X -> The.X
I've gotten rid of the necessity of typing "Scene." when reference one of the global resources. To do this, "The.cs" was created which requires every field of Scene to be wrapped in The.

I decided against this before, but the attractiveness of a much shorter and simpler way of referencing these resources (as well as more linguistically pleasant e.g. The.Users, The.SystemMap) moved me to create a script that automates this task.

I'm not sure I'll end up using the script long term, or just manually add the new lines, but either way I think its worth the 6 less boilerplate characters.