Moire reduction, False pixels fix, Misc visual improvement
This commit adds a new step the rendering process : after rendering the scene, it goes over the resulting texture with a gaussian-ish smoothing kernel.

To accomplish this, the previous final camera now has a post-processing script that writes to another texture, which is then captured by a third camera.

In addition, I've adjusted the resolution of the monitor screen's texture to be a little below the actual output resolution. The combination of these changes substantially reduces the appearance of moire.

Another visual pain point prior to this commit was that bright rows of pixels appeared to flow into each other, but grayer/darker rows did not; they instead appeared very noisily as separate pixels, but much worse was the fact that these apparent pixels were actually the point at which two separate pixels ran into each other, creating a spot halfway between that was much brighter than it was supposed to be (which also means the colors weren't correct).

To fix this, I select the maximum allowed intensity as the intensity of the underlying signal. E.g. if the underlying signal is (0.6, 0.6, 0.6), the output resulting from that underlying signal cannot exceed that value. Because pixels can overlap, I determine this maximum by weighting the underlying signal intensities by the intensity they would ordinarily emit, and then divide by the total weight.

I also made a lot of little changes to make things look nicer.

I increased the saturation of the pixel texture to improve visibility of colored fringes to pixels.

I added a "WhiteShift" field to Pixelize.shader, which attempts to reduce how pure the brightest colors appear (esp. pure green) by shifting them towards white. This was done because it appeared to me that the colors were too vivid to look like something you would see on an older screen that may struggle with trueness or be miscalibrated. With a little shifting, in my subjective opinion the realism is improved. Shifting is accomplished by basically computing the squared variance from the mean channel value. E.g. (0.9, 0, 0) => mean of 0.3.

Finally, adjusted some existing Pixelize.shader and Diverge.shader parameters.