UnityEditor.EditorApplication -> Application, UNITY_EDITOR macro in ReadOnlyAttribute.cs
This commit makes changes necessary to allow the project to be built into a standalone executable.

Use of UnityEditor.EditorApplication prevents building because this does not exist outside the editor. Simply referencing "Application" instead fixes this.

ReadOnlyAttribute uses UnityEditor.PropertyDrawer, which, again, does not exist outside the editor environment. To fix this, I've added a UNITY_EDITOR macro that removes the offending class when compiling for standalone.