ItemCollectionPanel, InventoryPage bugfix
ItemCollectionPanel had a problem where it tried to populate before its .Items were initialized, so I moved this initialization to InventoryPanel's Start() method. However, this meant InventoryPage was accessing .Storage.Items in the same step as the classes that initialize .Storage, merely moving the problem somewhere else.

The resolution was simply to make both of these classes check continuously in Update() whether the relevant field has been initialized, and only then using it.