Meeting.cs
Often, User.Scripts have very computationally expensive decision making. For example, the Manufacturer class has to estimate the ROI on every Machine available at every station the Manufacturer does business, then buy a small amount of the best machine, and repeat this process over and over until it has spent its investment capital.

The simplest way to reduce these costs is to simply run the decision making algorithm less often. I decided that the most intuitive way to accomplish this within the context of a game is to have these decision be made on a regular schedule in-world. Meaning, f.e. once every in-world month, a Manufacturer buys/sells machines and sells the products made by these machines.

To formalize this approach, which was used by both Manufacturer and People (with separate implementations), I created  "Meeting".

The word, in this context, does not refer to a specific instance - e.g. when people met on XX/XX/20XX, but rather the concept of a regular gathering for a specific purpose (Think "GDC", "The State of The Union"). The Meeting doesn't cease to exist after people go back to their desks. To avoid confusion, specific instances of a Meeting are called "sessions".