Item.IsEquivalent, Item.EquivalenceScript, ItemDatabase
In order to formalize item fungibility as well as make it programmable, I've got rid of the ad-hoc comparison by Item.Name, and replaced it with Item.IsEquivalent(Item other). Now, everything that wishes to compare items for stackability/fungibility uses this method. By default, Item only considers two Items equivalent if they are reference equal. In order to change this behavior, you must add Item.EquivalenceScripts to the Item, which can perform arbitrary comparisons. The script "Commodity" does exactly this (currently continuing the comparison-by-name implementation) and is intended to be the default script for this purpose.

Because of this change, items can no longer be identified by name, so now they are identified by example. For instance, in Market, if you want to know how many widgets are available, you used to call GetTotalSupply("Widgets"), now you call the same method but you instead pass an example widget instead. Because you generally can't wish to buy something that isn't already for sale, (how would you know it exists?), this doesn't usually change the implementation of most code other than changing Method(item.Name) to Method(item).

There are two exceptions. People.ItemNeeds and Recipe are both designed to be edited within Unity, and the editor-friendly dictionary asset I'm using doesn't support prefab references as keys. Therefore these classes use string keys. In order to resolve this problem, I use another new class in this commit, "ItemDatabase".

ItemDatabase's main purpose is to provide permanent examples of fungible Items (referred to as "samples" within the context of ItemDatabase). When you ask it for a sample of Item, it creates a copy of it and stores it inside itself throughout the run of the application, ensuring that this sample is also valid. If an equivalent sample already exists, it simply returns that. For Items that aren't fungible (referred to as "unique" Items), it simply stores a reference. Unique Items should preferably have a permanent lifetime to avoid their samples from becoming invalid.

Using the ItemDatabase, samples can be found from string identifiers for the purposes or in-editor dictionaries. Items identified in this way should only be items that don't have any special fungibility rules.